% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{gauge}
\alias{gauge}
\title{Gauge Widget}
\usage{
gauge(id, value = 0, options = NULL, wrap = TRUE)
}
\arguments{
\item{id}{a character string to specify the unique id for the widget}

\item{value}{the progress value to display (should be between 0 and 100)}

\item{options}{an optional list of parameters to tune the widget}

\item{wrap}{whether or not the widget should be wrapped into a div container}
}
\value{
an HTML widget
}
\description{
An animated SVG gauge widget
}
\details{
\code{id} is used to compute ids for children nodes of the svg widget, pretty much
like the namespace id in Shiny modules.
To ensure uniqueness when used in Shiny modules, the id of the module instance
should be used as a prefix \verb{id = paste(id, "gauge", sep = "-)}.

The div container used when \code{wrap = TRUE} comes with basic style (width: 300px,
height: 300px, margin: 20px) to avoid the widget to occupy the whole page.
If you wish to specify other values or options, set \code{wrap = FALSE} and encapsulate
the widget in your own div.
}
\examples{
\dontrun{
gauge(id = "foo")
}
}
