% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{Progress Widget}
\usage{
progress(id, value = 0, options = NULL, wrap = TRUE)
}
\arguments{
\item{id}{a character string to specify the unique id for the widget}

\item{value}{the progress value to display (should be between 0 and 100)}

\item{options}{an optional list of parameters to tune the widget}

\item{wrap}{whether or not the widget should be wrapped into a div container}
}
\value{
an HTML widget
}
\description{
Progress Widget
}
\examples{
\dontrun{
progress()
}
}
