% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_update.R
\name{progress_update}
\alias{progress_update}
\title{Update Progress Widget}
\usage{
progress_update(id, value, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{the widget to update}

\item{value}{the new progress value to display}

\item{session}{the session of object of the server function}
}
\description{
Check if the required dependency needs to be inserted on UI side
and update the progress widget.
}
\examples{
\dontrun{
progress_update(value = 52)
}
}
