% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge_update.R
\name{gauge_update}
\alias{gauge_update}
\title{Update Gauge Widget}
\usage{
gauge_update(value, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{value}{the new progress value to display}

\item{session}{the session of object of the server function}
}
\description{
Check if the required dependency needs to be inserted on UI side
and update the gauge widget.
}
\examples{
\dontrun{
gauge_update(value = 52)
}
}
